# Conflicts:
#	src/main/java/org/example/MembershipRepository.java
#	src/main/java/org/example/PerkRepository.java
#	src/main/java/org/example/RatingRepository.java
#	src/main/java/org/example/User.java
#	src/main/java/org/example/UserRepository.java

+++ src/main/java/org/example/Rating.java
@@ -7,7 +7,8 @@ import jakarta.persistence.ManyToOne;
 @Entity
 public class Rating {
     @Id
-    private Long rid;
+
+    private Integer rid;
 
     @ManyToOne
     @JoinColumn(name = "uid")
@@ -30,7 +31,7 @@ public class Rating {
     }
 
     public long getRid() { return rid; }
-    public void setRid(long rid) {
+    public void setRid(Integer rid) {
         this.rid = rid;
     }
